Refactor article components and update index page structure
- Remove display subdirectory and reorganize article components
- Update index page to use new LinksTopView component
- Remove dummy articles and featured article sections
- Simplify component imports and file structure
