+++ feature/article/components/cards/FeaturedCard.tsx
