+++ feature/article/components/index.tsx
@@ -1,7 +1,10 @@
 // Cards
-export { HorizontalCard } from "./display/cards/HorizontalCard";
-export { FeaturedArticleCard } from "./display/cards/FeaturedCard";
+export { HorizontalCard } from "./cards/HorizontalCard";
+export { FeaturedArticleCard } from "./cards/FeaturedCard";
 
 // Lists
-export { FeaturedArticleList } from "./display/lists/FeaturedList";
-export { ArticleFlatList } from "./display/lists/ArticleFlatList";
+export { FeaturedArticleList } from "./lists/FeaturedList";
+export { ArticleFlatList } from "./lists/ArticleFlatList";
+
+// View
+export { LinksTopView } from "./LinksTopView";
